♻️ Moved buildRecipients out of the Email model
buildRecipients joins the rest of the send pipeline in EmailSendService.
The moved code is byte-for-byte identical apart from taking the email as
a parameter instead of `this`.

This was the Email model's last read of the platform, and its last use of
EmailBuilder. @stamhoofd/models no longer resolves a platform anywhere in
Email.ts, and the only remaining EmailBuilder importer inside the package
is Organization.updateDNSRecords.
